Layer tree view (#1748)
* initial version of Layer Tree Widget (LC_LayerTreeWidget)

* better naming of tree view grid delegate class to avoid possible merge issues

* merged cmake file

* fixed merge issue that prevents build

---------

+++ librecad/src/ui/qg_graphicview.cpp
@@ -787,7 +787,7 @@ void QG_GraphicView::wheelEvent(QWheelEvent *e) {
 #if (QT_VERSION >= QT_VERSION_CHECK(5, 15, 0))
             e->position(),
 #else
-            QPointF{e->x(), e->y},
+            QPointF{static_cast<qreal>(e->x()), static_cast<qreal>(e->y())},
 #endif
             Qt::NoButton, Qt::NoButton, Qt::NoModifier
     };
